Bug fix: user's personal rating now correctly displayed on reload

+++ backend/server.js
@@ -172,46 +172,56 @@ app
 
     const userRecipeRating = `SELECT rating, recipe_id FROM recipe_rating WHERE recipe_id = $1 AND user_id = $2`
 
-    const recipe = (await client.queryObject(userRecipeRating, recipeId, currentUser.id)).rows
+    const [recipe] = (await client.queryObject(userRecipeRating, recipeId, currentUser.id)).rows
 
-    console.log( recipe )
-
-    if(recipe.length === 0){
-      server.json({ rating:  0, recipe_id: recipeId})
-    }else{
-      server.json(recipe)
+    if (currentUser) {
+      if(recipe && recipe.rating) {
+        //All good: recipe rated
+        server.json({ response:'success', recipe })
+      } else {
+        //All good: recipe not rated
+        server.json({response:'success', recipe: { rating:  0, recipe_id: recipeId }})
+      }
+    } else {
+      //Bad Credentials
+      await server.json({ response: 'unauthorized', recipe: {} })
     }
+
   })
 
 
   //--------------------------- Post Rating ------------------------//
 
   .post('/recipe/rating', async (server) => {
+    const { sessionId } = server.cookies
+    const currentUser = await getCurrentUser(sessionId)
 
-    if(!server.cookies.sessionId){
+    if (currentUser) {
+      const { rating, recipeId } = await server.body
+
+      //Is sequential id numbering a security risk
+      const user = await getCurrentUser(server.cookies.sessionId)
+      //Delete instances with same recipe id and user.id
+      const deleteOldRating = `
+        DELETE FROM recipe_rating
+          WHERE user_id = $1 AND recipe_id = $2;`
+
+      const insertNewRating = `INSERT INTO recipe_rating 
+          (rating, created_at, updated_at, recipe_id, user_id)
+        VALUES 
+          ($1, NOW(), NOW(), $2, $3)
+          RETURNING rating;`
+
+        await client.queryObject(deleteOldRating, user.id, recipeId)
+        
+        const [ ratingResponse ] = (await client.queryObject(insertNewRating, rating, recipeId, user.id)).rows
+
+      //All good: return user's rating
+      server.json({ rating: ratingResponse.rating })
+    } else {
+      //User is unauthrized to rate
       server.json({message: 'You need to be a registered user to rate.'})
     }
-
-    const { rating, recipeId } = await server.body
-
-    //Is sequential id numbering a security risk
-    const user = await getCurrentUser(server.cookies.sessionId)
-    //Delete instances with same recipe id and user.id
-    const deleteOldRating = `
-      DELETE FROM recipe_rating
-        WHERE user_id = $1 AND recipe_id = $2;`
-
-    const insertNewRating = `INSERT INTO recipe_rating 
-        (rating, created_at, updated_at, recipe_id, user_id)
-      VALUES 
-        ($1, NOW(), NOW(), $2, $3)
-        RETURNING rating;`
-
-      await client.queryObject(deleteOldRating, user.id, recipeId)
-      
-      const [ ratingResponse ] = (await client.queryObject(insertNewRating, rating, recipeId, user.id)).rows
-
-    server.json({ rating: ratingResponse.rating })
 
   })
